📖 Add new parser test doc, move uid_mapping test
