+++ test/net/imap/test_imap_response_data.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+require "net/imap"
+require "test/unit"
+
+class IMAPResponseDataTest < Test::Unit::TestCase
+
+  def setup
+    @do_not_reverse_lookup = Socket.do_not_reverse_lookup
+    Socket.do_not_reverse_lookup = true
+  end
+
+  def teardown
+    Socket.do_not_reverse_lookup = @do_not_reverse_lookup
+  end
+
+  def test_uidplus_copyuid__uid_mapping
+    parser = Net::IMAP::ResponseParser.new
+    response = parser.parse(
+      "A004 OK [copyUID 9999 20:19,500:495 92:97,101:100] Done\r\n"
+    )
+    code = response.data.code
+    assert_equal(
+      {
+         19 =>  92,
+         20 =>  93,
+        495 =>  94,
+        496 =>  95,
+        497 =>  96,
+        498 =>  97,
+        499 => 100,
+        500 => 101,
+      },
+      code.data.uid_mapping
+    )
+  end
+
+end
